+++ include/class-email-log-list-table.php
@@ -152,6 +152,15 @@ class Email_Log_List_Table extends WP_List_Table {
 			__( 'Delete', 'email-log' )
 		);
 
+		/**
+		 * This filter can be used to modify the list of row actions that are displayed.
+		 *
+		 * @since 1.8
+		 * @param array $actions List of actions.
+		 * @param object $item The current log item.
+		 */
+		$actions = apply_filters( 'el_row_actions', $actions, $item );
+
 		return sprintf( '%1$s <span style="color:silver">(id:%2$s)</span>%3$s',
 			/*$1%s*/ $email_date,
 			/*$2%s*/ $item->id,